chore: move activity packages to feature
