+++ app/src/main/java/com/team22/soundary/feature/profile/ProfileActivity.kt
@@ -1,0 +1,4 @@
+package com.team22.soundary.feature.main.profile
+
+class ProfileActivity {
+}
