+++ app/src/main/java/com/team22/soundary/feature/share/ShareFriendActivity.kt
@@ -1,12 +1,12 @@
-package com.team22.soundary.search
+package com.team22.soundary.feature.share;
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.team22.soundary.R
 
-class FriendSearchActivity : AppCompatActivity() {
+class ShareFriendActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_friend_search)
+        setContentView(R.layout.activity_share_friend)
     }
 }
