--- app/src/main/java/com/team22/soundary/profile/ProfileActivity.kt
@@ -1,4 +1,0 @@
-package com.team22.soundary.profile
-
-class ProfileActivity {
-}
