+++ app/src/main/res/layout/activity_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".profile.ProfileActivity">
+    tools:context=".feature.main.profile.ProfileActivity">
 
     <LinearLayout
         android:id="@+id/vertical_container"
